Get Phylogeny of Angiosperms

Retreive seed plant phylogenies described in Smith & Brown 2018 from FePhyFoFum in github

Usage

get_phy_angio(phy_choice)

Arguments

phy_choice

Character string specifying which seed plant phylogeny to download. Must be one of:

"ALLMB": GenBank and Open Tree of Life taxa with a backbone from Magallón et al. (2015)

"ALLOTB": GenBank and Open Tree of Life taxa with a backbone from Open Tree of Life version 9.1

"GBMB": GenBank taxa with a backbone from Magallón et al. (2015)

"GBOTB": GenBank taxa with a backbone from Open Tree of Life version 9.1

Value

A rooted phylogenetic tree as a phylo object to be used in comparative methods

Details

The function downloads a pre-constructed phylogenetic tree from the big_seed_plant_trees GitHub repository, unzips it to a temporary directory, and reads the .tre file corresponding to the selected phy_choice.

References

Smith S, Brown J (2018). “Constructing a broadly inclusive seed plant phylogeny.” American Journal of Botany. doi:10.1002/ajb2.1019 . Kleyer M, et al. (2008). “The LEDA Traitbase: a database of life-history traits of the Northwest European Flora.” Journal of Ecology. doi:10.1111/j.1365-2745.2008.01430.x . Magallón S, et al. (2015). “A metacalibrated time-tree documents the early rise of flowering plant phylogenetic diversity.” New Phytologist. doi:10.1111/nph.13264 .

Author

Alivia G Nytko, anytko@vols.utk.edu

Examples

if (FALSE) { # \dontrun{
ALLMB_phy <- get_phy_angio(phy_choice = "ALLMB")

ALLOTB_phy <- get_phy_angio(phy_choice = "ALLOTB")

GMBM_phy <- get_phy_angio(phy_choice = "GBMB")

GBOTB_phy <- get_phy_angio(phy_choice = "GBOTB")
} # }